Actor should be set even if not authenticated
Previously the user was set as the actor ONLY if the user is
authenticated when middleware is processed. Now, authentication is
checked in the `AuditlogMiddleware.set_actor` method, instead of
skipping that if a user is not authenticated when the middleware's
`process_request` method is called. This occurs when using Django Rest
Framework (DRF), as it sets a request's user to the authenticated user
after all middleware has been processed. By checking the user object
after the DRF authenticates, through a `SimpleLazyObject` (see:
https://crondev.blog/2018/05/13/django-middlewares-with-rest-framework/),
we can properly set the actor when using this with DRF.